feat(webapp): allow a fractional ck vtime quantum
The vtime Lua serves weighted fair-queue tags and handles a sub-1 quantum (a
finer serve granularity), and ckVtimeRegistration.test.ts exercises it at 0.3,
but RUN_ENGINE_CK_VTIME_QUANTUM was schema-constrained to integers. Drop the
.int() so the env no longer blocks a capability the engine already has. Default
stays 1 (the weight hook is fixed at 1 today).

Flagged by Devin on #4367.

+++ apps/webapp/app/env.server.ts
@@ -1224,7 +1224,10 @@ const EnvironmentSchema = z
     // Fair (virtual-time) ordering across concurrency-key variants of a base queue.
     // Off by default; when off the run queue behaves exactly as before.
     RUN_ENGINE_CK_VTIME_SCHEDULING_ENABLED: BoolEnv.default(false),
-    RUN_ENGINE_CK_VTIME_QUANTUM: z.coerce.number().int().positive().default(1),
+    // Fractional allowed: the vtime Lua serves weighted fair-queue tags, so a sub-1 quantum
+    // is a valid finer serve granularity. The weight hook is fixed at 1 today, so 1 stays the
+    // default, but the schema no longer blocks the capability the engine already has.
+    RUN_ENGINE_CK_VTIME_QUANTUM: z.coerce.number().positive().default(1),
     RUN_ENGINE_CK_VTIME_WINDOW_MULTIPLIER: z.coerce.number().int().positive().default(3),
     RUN_ENGINE_CK_VTIME_STATE_TTL_SECONDS: z.coerce.number().int().positive().default(86400),
